requirements.txt
boto3==1.18.59
psycopg2-binary==2.9.1